resolve conflicts - move stuff to conftest
