+++ tests/test_cyaccess.py
@@ -1,7 +1,5 @@
 # test direct use of cyccess (c++/cython extension)
 
-import pathlib
-
 import pandas as pd
 import numpy as np
 import pytest
@@ -11,7 +9,7 @@ from pandarm.cyaccess import cyaccess
 
 @pytest.fixture(scope="module")
 def nodes_and_edges(request):
-    store = pd.HDFStore(pathlib.Path(__file__).parent / "osm_sample.h5", "r")
+    store = pd.HDFStore(pytest.h5_osm_sample, "r")
     nodes = store.nodes
     edges = store.edges[["from", "to"]]
     edge_weights = store.edges[["weight"]]
