+++ tests/test_pandash5.py
@@ -1,6 +1,3 @@
-import pathlib
-import tempfile
-
 import pandas as pd
 import pytest
 from pandas.testing import assert_frame_equal, assert_series_equal
@@ -41,9 +38,7 @@ def two_way():
 @pytest.fixture(scope="module")
 def network(nodes, edges, edge_weights, two_way):
     with pytest.no_crs_warning:
-        return Network(
-            nodes["x"], nodes["y"], edges["from"], edges["to"], edge_weights, two_way
-        )
+        return Network(nodes["x"], nodes["y"], edges["from"], edges["to"], edge_weights, two_way)
 
 
 @pytest.fixture(scope="module")
@@ -54,19 +49,6 @@ def edges_df(edges, edge_weights):
 @pytest.fixture(scope="module")
 def rm_nodes():
     return [0, 7, 6]
-
-
-@pytest.fixture
-def tmpfile(request):
-    fname = pathlib.Path(tempfile.NamedTemporaryFile().name)
-
-    def cleanup():
-        if fname.exists():
-            fname.unlink()
-
-    request.addfinalizer(cleanup)
-
-    return fname
 
 
 def test_remove_nodes(network, rm_nodes):
